Move `servers/rest.Route{,s}` into its own package
... `servers/rest/route`.

This will help avoid recursive imports, that would happen when Rest is
moved out of gdctx.
