+++ commands/peers/commands.go
@@ -2,7 +2,7 @@
 package peercommands
 
 import (
-	"github.com/gluster/glusterd2/servers/rest"
+	"github.com/gluster/glusterd2/servers/rest/route"
 )
 
 // Command is a holding struct used to implement the GlusterD Command interface
@@ -10,30 +10,30 @@ type Command struct {
 }
 
 // Routes returns command routes. Required for the Command interface.
-func (c *Command) Routes() rest.Routes {
-	return rest.Routes{
-		rest.Route{
+func (c *Command) Routes() route.Routes {
+	return route.Routes{
+		route.Route{
 			Name:        "GetPeer",
 			Method:      "GET",
 			Pattern:     "/peers/{peerid}",
 			Version:     1,
 			HandlerFunc: getPeerHandler,
 		},
-		rest.Route{
+		route.Route{
 			Name:        "GetPeers",
 			Method:      "GET",
 			Pattern:     "/peers",
 			Version:     1,
 			HandlerFunc: getPeersHandler,
 		},
-		rest.Route{
+		route.Route{
 			Name:        "DeletePeer",
 			Method:      "DELETE",
 			Pattern:     "/peers/{peerid}",
 			Version:     1,
 			HandlerFunc: deletePeerHandler,
 		},
-		rest.Route{
+		route.Route{
 			Name:        "AddPeer",
 			Method:      "POST",
 			Pattern:     "/peers",
