+++ commands/version/commands.go
@@ -2,7 +2,7 @@
 package versioncommands
 
 import (
-	"github.com/gluster/glusterd2/servers/rest"
+	"github.com/gluster/glusterd2/servers/rest/route"
 )
 
 // Command is a holding struct used to implement the GlusterD Command interface
@@ -10,9 +10,9 @@ type Command struct {
 }
 
 // Routes returns command routes. Required for the Command interface.
-func (c *Command) Routes() rest.Routes {
-	return rest.Routes{
-		rest.Route{
+func (c *Command) Routes() route.Routes {
+	return route.Routes{
+		route.Route{
 			Name:        "GetVersion",
 			Method:      "GET",
 			Pattern:     "/version",
