+++ servers/rest/route/route.go
@@ -1,0 +1,20 @@
+// Package route implements a Route type to be used by GD2 rest API to define their routes
+package route
+
+import (
+	"net/http"
+)
+
+// Route models a route to be set on the GlusterD Rest server
+// This route style comes from the tutorial on
+// http://thenewstack.io/make-a-restful-json-api-go/
+type Route struct {
+	Name        string
+	Method      string
+	Pattern     string
+	Version     int
+	HandlerFunc http.HandlerFunc
+}
+
+// Routes is a table of many Route's
+type Routes []Route
